De-spaghettified cutscene solution
Cutscenes are controlled by timelines. For everything that can be stationary while the game waits for player input to continue, normal animation tracks can be used. For everything that has a looping animation while the game waits for player input, a signal track must be used. The signal track will tell each animation to start playing. This does mean that animation blending isn't supported for character animation, but if anything, that's a plus, because it fits the style. However, this does mean shenanigans and lots of animator spaghetti.