Rewrote item system to work with save/load system, moved start menu into main scene, UI design, cutscene work, and combined the menu scripts
Yes, this is a behemoth of a commit. Items were scripts that inherited (indirectly) from monobehavior, making them regular components. You couldn't create them at runtime unless you put them on objects (like how you can't have a collider without an object), so when the save system tried to create *EVERY* item in a dictionary to store them in a list to keep track of the items the player picked up, Unity decided to file a restraining order and take an extended vacation. Rather than bandaiding the save system in 30 minutes like a sane person would, I took it as an opportunity to refactor the items using scriptable objects. But now, both the save system and item system play nice, Unity quit being dramatic, and I spent 3 hours refactoring a system that didn't really need to be because I felt like it :)
At least it was fun