A proper finish
I needed to grab screenshots and gifs for my portfolio, so I thought that was a good excuse to come back and fix the last-second issues I created. Turns out crunching for a deadline causes burnout, and that meant I couldn't touch the project at all until now. Made a couple of new minor improvements, implemented a few small things that were supposed to be working but got broken, and finally fixed those damn merge conflicts.